* Support for version 1 mvhd
git-svn-id: svn://svn.code.sf.net/p/gmerlin/code/trunk@811 214d8ac5-8513-4f59-80f8-2f39b674a24a

+++ gmerlin_avdecoder/include/qt.h
@@ -616,10 +616,10 @@ typedef struct
   qt_atom_header_t h;
   int version;
   uint32_t flags;
-  uint32_t creation_time;
-  uint32_t modification_time;
+  uint64_t creation_time;
+  uint64_t modification_time;
   uint32_t time_scale;
-  uint32_t duration;
+  uint64_t duration;
   float preferred_rate;
   float preferred_volume;
   uint8_t reserved[10];
